refactor formatValueByType
