+++ packages/utils/package.json
@@ -26,6 +26,7 @@
     "moment": "^2.24.0",
     "moment-timezone": "^0.5.25",
     "node-fetch": "^1.7.3",
+    "numeral": "^2.0.6",
     "validator": "^8.2.0",
     "winston": "^3.3.3",
     "xlsx": "^0.10.9"
