+++ packages/utils/src/constant.js
@@ -4,5 +4,13 @@
  */
 
 export const VALUE_TYPES = {
+  BOOLEAN: 'boolean',
+  CURRENCY: 'currency',
+  FRACTION: 'fraction',
+  PERCENTAGE: 'percentage',
+  FRACTION_AND_PERCENTAGE: 'fractionAndPercentage',
   NUMBER_AND_PERCENTAGE: 'numberAndPercentage',
+  TEXT: 'text',
+  NUMBER: 'number',
+  ONE_DECIMAL_PLACE: 'oneDecimalPlace',
 };
